+++ tests/testthat/test-targets.R
@@ -52,10 +52,12 @@ test_that("Workflow works", {
         ~organization_name, ~project_name,
         "北京师范大学", "4.19-4.20夜晚睡眠test"
       )
-      tar_prep_iquizoo(params)
+      tar_prep_iquizoo(params, combine = "scores")
     })
     targets::tar_make(reporter = "silent", callr_function = NULL)
     expect_snapshot_value(targets::tar_objects(), style = "json2")
+    expect_snapshot_value(targets::tar_read(users), style = "json2")
+    expect_snapshot_value(targets::tar_read(scores), style = "json2")
   })
 })
 
@@ -147,5 +149,25 @@ test_that("Ensure project date is used", {
     })
     targets::tar_make(reporter = "silent", callr_function = NULL)
     nrow(targets::tar_read(scores)) |> expect_gt(0)
+  })
+})
+
+test_that("Ensure `check_progress = FALSE` work", {
+  targets::tar_dir({
+    targets::tar_script({
+      library(targets)
+      tar_prep_iquizoo(
+        contents = data.frame(
+          project_id = bit64::as.integer64(132121231360389),
+          game_id = bit64::as.integer64(268008982646879)
+        ),
+        what = "scores",
+        check_progress = FALSE
+      )
+    })
+    targets::tar_make(reporter = "silent", callr_function = NULL)
+    expect_true(
+      all(!startsWith(targets::tar_objects(), "progress_hash"))
+    )
   })
 })